% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.t.test.plus.R
\name{power.t.test.plus}
\alias{power.t.test.plus}
\title{Enhanced output for Power calculations for one and two sample t tests}
\usage{
power.t.test.plus(..., complete_rate = 0.8)
}
\arguments{
\item{...}{options in 'power.t.test'}

\item{complete_rate}{the complete rate of the clinical trail}
}
\description{
All options is same as the 'power.t.test' We also generate Half width for 90% and 95% Confidence interval.
}
\examples{
power.t.test.plus(n = 100, delta = 0.5)
}

